Refactor favorites handling: implement custom validation error struct for improved error handling in favorite operations
